Update Board.py
Commented out call to test(); it is no longer valuable due to Puzzle.py handling Board creation and population in a more concrete fashion. The Block class now returns ints. Serious changes to all directional "canMove" functions to get them to work properly within the bounds of the board were made. RemoveBlock() works again. Block class changed to account for the order in which height and width are read from the file. Changes were made to AddBlock() (as well as other affected functions) to account for this. The Board class now initializes the width and height of the tray. The Block class now contains an ID attribute and a getID() method. The ID is set in Puzzle.py. Changes were made to AddBlock() and RemoveBlock() to account for elements of the Block class having IDs. Due to the number of changes made, the code may still be broken in certain spots. Expect this to require some debugging.